requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.2
Django==4.2.6
django-admin-list-filter-dropdown==1.0.3
django-ckeditor-5==0.2.10
djangorestframework==3.14.0
idna==3.4
Pillow==10.1.0
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.1.0
